requirements.txt
flask==3.0.2
python-dotenv==1.0.1
requests==2.31.0
flask-cors==4.0.0
circlemind==0.0.5.post2 